requirements.txt
mutagen==1.47.0
numpy==2.2.6
opencv-python==4.12.0.88
pillow==12.0.0
pydub==0.25.1
